arch doc and deploy update

+++ start-services.sh
@@ -1,17 +1,6 @@
 #!/bin/bash
 
-source ~/env/env.sh
-cp ~/env/security.keystore.cloud ~/app/scripts/
-cp ~/env/TC.cloud.ldap.keystore ~/app/scripts/
-
-echo Creating SQS queues...
-sleep 2
-java -cp "/home/apps/app/lib/jars/*" -Darena.sqs-endpoint="$SQS_URL" com.topcoder.arena.util.sqs.LocalSqsSetup "$SQS_URL" devArenaCode- compile admin-test listener-results proc-results mm-test practice srm-test
-
-# also set dummy aws credentials
-SQS_AWS_OPTS="-Darena.sqs-endpoint=$SQS_URL -Darena.env-prefix=dev -Daws.accessKeyId=x -Daws.secretKey=x"
-export JAVA_OPTS="$SQS_AWS_OPTS -DconfigurationProvider.class=com.topcoder.farm.controller.configuration.XMLConfigurationProvider -Dconfiguration.xml.url=file:///home/apps/app/controller/config.xml"
-export JBOSS_JAVA_OPTS="$JBOSS_JAVA_OPTS -Djavax.net.ssl.trustStore=TC.cloud.ldap.keystore"
+source ~/env/arena-env.sh
 
 export JAVA_VER=$(javap -verbose java.lang.Object | grep "major version" | cut -d " " -f5)
 
@@ -20,43 +9,72 @@ if [ "$JAVA_VER" -ge 55 ] ; then
   export CP=$CP:~/app/lib/corba/*
 fi
 
-cd ~/app/scripts
+CMD=usage
+if [[ $1 != "" ]] ; then
+	CMD=$1
+	shift
+fi
 
-# start jboss
-echo Starting JBoss...
-./jboss.sh
+if [ "$CMD" = "app" ] ; then
+	export JAVA_OPTS="$SQS_AWS_OPTS -DconfigurationProvider.class=com.topcoder.farm.controller.configuration.XMLConfigurationProvider -Dconfiguration.xml.url=file:///home/apps/app/controller/config.xml"
+	export JBOSS_JAVA_OPTS="$JBOSS_JAVA_OPTS -Djavax.net.ssl.trustStore=TC.cloud.ldap.keystore"
 
-echo Waiting for startup...
-sleep 30
+	cd ~/app/scripts
 
-echo Starting main listener...
-./runMainListener.sh
+	echo Starting JBoss...
+	./jboss.sh
 
-echo Waiting for listener to initialize...
-sleep 30
+	echo Waiting $JBOSS_WAIT_TIME seconds for JBoss startup...
+	sleep $JBOSS_WAIT_TIME
 
-echo Starting admin listener...
-./runAdminListener.sh
+	echo Starting Main Listener...
+	./runMainListener.sh
 
-echo Starting mpsqas listener...
-./mpsqasListener.sh start 5037
+	echo Waiting $MAIN_LISTENRER_WAIT_TIME seconds for Main Listener startup...
+	sleep $MAIN_LISTENRER_WAIT_TIME
 
-echo Starting web socket listener...
-./runWebSocketListener.sh
+	echo Starting Admin Listener...
+	./runAdminListener.sh
 
-echo Starting Farm Processor...
-rm -rf ~/processor/work
-mkdir ~/processor/work
-rm -rf ~/processor/cache
-mkdir ~/processor/cache
-chmod +x ~/processor/deploy/app/cpp/timeout/timeout
+	echo Waiting for Admin Listener startup...
+	sleep 5
 
-export JAVA_OPTS="-DcppArguments=-std=c++17 $SQS_AWS_OPTS"
+	echo Starting MPSQAS Listener...
+	./mpsqasListener.sh start 5037
 
-cd ~/processor/deploy/bin
-./processor.sh PR-LX
+	echo Waiting for MPSQAS Listener startup...
+	sleep 5
 
-echo Startup complete
+	echo Starting WebSocket Listener...
+	./runWebSocketListener.sh
+
+	echo Waiting for WebSocket Listener startup...
+	sleep 5
+
+	echo Arena app startup complete
+elif [ "$CMD" = "processor" ] ; then
+    echo Starting Farm Processor...
+	rm -rf ~/processor/work
+	mkdir ~/processor/work
+	rm -rf ~/processor/cache
+	mkdir ~/processor/cache
+	chmod +x ~/processor/deploy/app/cpp/timeout/timeout
+
+	export JAVA_OPTS="-DcppArguments=-std=c++17 $SQS_AWS_OPTS $PROCESSOR_OPTS"
+
+	cd ~/processor/deploy/bin
+	./processor.sh PR-LX
+
+	echo Waiting for Farm Processor startup...
+	sleep 5
+
+	echo Arena processor startup complete
+else
+    echo "Usage:"
+    echo "start-services.sh (app|processor)"
+    echo "	app       - start JBoss and Listeners in the background"
+    echo "	processor - start Farm Processor in the background"
+fi
 
 sleep infinity
